% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Categorizes metabolites to categories}
\usage{
categorize(df)
}
\arguments{
\item{df}{A data frame containing a `HMDB` column.}
}
\value{
A data frame with columns `Class`, `SubClass` and `Category`.
}
\description{
The function matches the HMDB ID of a metabolite to HMDB classes and sub-classes and a manually curated category.
It is recommended that the dataframe is first standardized using the `standardize` function.
}
\examples{
data <-  data.frame(
Name = c("Glucose", "Tyrosine", "Valine"),
Compound = c("D-Glucose", "L-Tyrosine", "L-Valine"),
HMDB = c("HMDB0000122", "HMDB0000158", "HMDB0000883"),
KEGG = c("C00031", "C00082", "C00183"),
Concentration = c(4500, 3000, 2000)
)

categorize(data)
}
