% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_comparison.R
\name{load_comparison}
\alias{load_comparison}
\title{Loads comparison data sets}
\usage{
load_comparison(df, medium, defined = TRUE)
}
\arguments{
\item{df}{An optional dataframe containing sample metabolite composition to append the commercial media composition to for comparison.}

\item{medium}{An optional vector to only load one or more select media from a list of DMEM, aDMEM, DMEM.F12, aDMEM.F12, F12, F10, MEM, IMDM, RPMI, aRPMI, McCoy5A, NBM, StemPro, mTeSR, E6, E8, ECGM, EpiLife, N2B27, StemSpan, UC, PluriSTEM}

\item{defined}{If set to TRUE (by default), only fully defined commercial media are loaded}
}
\value{
A data frame with commercial media metabolite composition.
}
\description{
The function loads media composition of fully and partially defined commercial media
}
\examples{
data <-  data.frame(
Name = c("Glucose", "Tyrosine", "Valine"),
Compound = c("D-Glucose", "L-Tyrosine", "L-Valine"),
HMDB = c("HMDB0000122", "HMDB0000158", "HMDB0000883"),
KEGG = c("C00031", "C00082", "C00183"),
Concentration = c(4500, 3000, 2000)
)

load_comparison(data, medium = c("aDMEM", "aRPMI", "F12"))
}
