% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barplot.R
\name{plot_barplot}
\alias{plot_barplot}
\title{Barplot to compare and cluster different cell culture media.}
\usage{
plot_barplot(
  data,
  category,
  n = 15,
  x_var = "Compound",
  y_var = "Concentration",
  fill_var = "Medium",
  x_lab = x_var,
  y_lab = y_var,
  fill_lab = fill_var,
  legend = "right",
  palette = "Dark2"
)
}
\arguments{
\item{data}{A data frame containing metabolomics/composition data of different cell culture media.}

\item{category}{Optional vector to filter to specified metabolite categories.}

\item{n}{The number of characters to retain when displaying metabolite names. Set to 15 by default.}

\item{x_var}{X axis variable. Set to `Compound` by default.}

\item{y_var}{Y axis variable. Set to `Concentration` by default.}

\item{fill_var}{Fill variable. Set to `Medium` by default.}

\item{x_lab}{X axis title. Set to the same value as x_var by default.}

\item{y_lab}{Y axis title. Set to the same value as y_var by default.}

\item{fill_lab}{Title for the fill legend. Set to the same value as fill_var by default.}

\item{legend}{Accepts the same values as ggplot. Set to `right` by default.}

\item{palette}{Color palette. Set to `Dark2` by default.}
}
\value{
A heatmap plot.
}
\description{
The function generates a barplot using the ggplot2 package
}
