% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap to compare and cluster different cell culture media.}
\usage{
plot_heatmap(
  data,
  category,
  n = 15,
  display_numbers = FALSE,
  clustering_method = "complete",
  cluster_rows = FALSE,
  cluster_cols = TRUE,
  scale = "row",
  fontsize_row = 8,
  fontsize_col = 8,
  ann_col = list(Category = c(Vitamins = "#8DEEEE", Carbohydrates = "#FFABA9",
    `Amino Acids` = "#007FFF", `Amino Acid Analogues and Peptides` = "#ee6363",
    `Metallic Compounds` = "#FFA54F", Lipids = "#AB82FF", Other = "#EEAEEE"))
)
}
\arguments{
\item{data}{A data frame containing metabolomics/composition data of different cell culture media.}

\item{category}{Optional vector to filter to specified metabolite categories.}

\item{n}{The number of characters to retain when displaying metabolite names. Set to 15 by default.}

\item{display_numbers}{Argument to display numbers in heatmap. Set to FALSE by default.}

\item{clustering_method}{Accepts the same values as pheatmap. Set to `complete` by default.}

\item{cluster_rows}{Accepts the same values as pheatmap. Set to FALSE by default.}

\item{cluster_cols}{Accepts the same values as pheatmap. Set to TRUE by default.}

\item{scale}{Accepts the same values as pheatmap. Set to `row` by default.}

\item{fontsize_row}{Accepts the same values as pheatmap. Set to 8 by default.}

\item{ann_col}{List of colors for each category.}

\item{fontsize_column}{Accepts the same values as pheatmap. Set to 8 by default.}
}
\value{
A heatmap plot.
}
\description{
The function generates a clustered heatmap using the pheatmap package
}
