% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize names in metabolomics data sets.}
\usage{
standardize(df)
}
\arguments{
\item{df}{A data frame containing a `Name` column.}
}
\value{
A standardized data frame with columns `Compound`, `HMDB` and `KEGG`.
}
\description{
The function matches the input name with a standardized name, HMDB ID and KEGG ID.
The function is a wrapper that uses MetaboAnalyst API.
}
\examples{
data <-  data.frame(
Name = c("Glucose", "Tyrosine", "Valine"),
Concentration = c(4500, 3000, 2000)
)

standardize(data)
}
